Added Version Check & Simple KillSwitch Now you can view the Latest Version of the Client (online) and see, If you need to update anything or not.
Maybe I will add a info screen for Updates and other stuff like that soon, maybe ...

Oh, and I have implemented a simple KillSwitch.
If I change it to a String (online document), the Client crashes. So only "real" DEVs can fix this, not everybody.
It's very simple, so don't blame meh.

Signed-off-by: realfantacoke <CrazyMemeCoke@gmail.com>