Reworked many stuff ARRAYLIST: The ArrayList is now sorted, thanks to Vinii for his method. I also had a method for that in a previous Client, but I couldn't find it anymore, so he helped me out. Also the Rect options are no longer available, from now on all Rects are displayed at the left side.
INVIS MODULE / MODE:
I put the Invis in here, because it's the only way to get it called every time the Client updates, w/o creating a entire new module.
I already know, how to solve that problem and to make it better (that will follow very soon)

MODULE INDICATIONS / VISUALIZATIONS:
Now the modules will display indications like the curTarget in Aura or the Speed mode in Speed.

Signed-off-by: realfantacoke <CrazyMemeCoke@gmail.com>